+++ api/Pipfile
@@ -45,7 +45,6 @@ python-jose = "*"
 uvarint = "*"
 django-ace = "*"
 async-timeout = "*"
-django-import-export = "*"
 
 [dev-packages]
 black = "*"
