+++ api/registry/admin.py
@@ -1,6 +1,5 @@
 from asgiref.sync import async_to_sync
 from django.contrib import admin, messages
-from import_export.forms import ModelResource
 from registry.api.schema import SubmitPassportPayload
 from registry.api.v1 import ahandle_submit_passport
 from registry.models import (
@@ -10,16 +9,8 @@ from registry.models import (
     Passport,
     Score,
     Stamp,
-    AddressListMember,
-    AddressList,
 )
 from scorer.scorer_admin import ScorerModelAdmin
-from import_export.admin import ImportMixin, ImportForm
-from django import forms
-from django.urls import path
-from django.shortcuts import render, redirect
-import csv
-import codecs
 
 
 @admin.action(
@@ -157,84 +148,3 @@ class GTCStakeEventAdmin(ScorerModelAdmin):
         "staker",
         "event_type",
     ]
-
-
-class AddressListMemberInline(admin.TabularInline):
-    model = AddressListMember
-    extra = 0
-
-
-class AddressListImportForm(ImportForm):
-    list = forms.ModelChoiceField(queryset=AddressList.objects.all(), required=True)
-
-
-class AddressListMemberResource(ModelResource):
-    def __init__(self, **kwargs):
-        super().__init__()
-        print(kwargs, "ALMR init - KWARGS!!!!!!")
-        self.list_id = kwargs.get("list_id")
-
-    class Meta:
-        model = AddressListMember
-
-
-class AddressListCsvImportForm(forms.Form):
-    csv_file = forms.FileField()
-    list = forms.ModelChoiceField(queryset=AddressList.objects.all(), required=True)
-
-
-@admin.register(AddressListMember)
-class AddressListMemberAdmin(ImportMixin, admin.ModelAdmin):
-    import_form_class = AddressListImportForm
-
-    def get_import_resource_kwargs(self, request, *args, **kwargs):
-        kwargs = super().get_resource_kwargs(request, *args, **kwargs)
-        print(kwargs, "girk - KWARGS!!!!!!")
-        # kwargs.update({"user": request.user})
-        return kwargs
-
-    def get_import_data_kwargs(self, **kwargs):
-        """
-        Prepare kwargs for import_data.
-        """
-        print(kwargs, "gidk - KWARGS!!!!!!")
-        form = kwargs.get("form", None)
-        if form and hasattr(form, "cleaned_data"):
-            print(form.cleaned_data, "gidk - FORM.CLEANED_DATA!!!!!!")
-            kwargs.update({"list_id": form.cleaned_data.get("list", None)})
-        return kwargs
-
-    def after_init_instance(self, instance, new, row, **kwargs):
-        print(kwargs, "KWARGS!!!!!!")
-        if "list_id" in kwargs:
-            instance.list_id = kwargs["list_id"]
-
-
-@admin.register(AddressList)
-class AddressListAdmin(ScorerModelAdmin):
-    list_display = ["name", "address_count"]
-    inlines = [AddressListMemberInline]
-    change_list_template = "registry/addresslist_changelist.html"
-
-    def address_count(self, obj):
-        return obj.addresses.count()
-
-    def get_urls(self):
-        return [
-            path("import-csv/", self.import_csv),
-        ] + super().get_urls()
-
-    def import_csv(self, request):
-        if request.method == "POST":
-            csv_file = request.FILES["csv_file"]
-            reader = csv.reader(codecs.iterdecode(csv_file, "utf-8"))
-            list_id = request.POST.get("list")
-            address_list = AddressList.objects.get(id=list_id)
-            for row in reader:
-                address = row[0].strip()
-                AddressListMember.objects.create(address=address, list=address_list)
-            self.message_user(request, "Your csv file has been imported")
-            return redirect("..")
-        form = AddressListCsvImportForm()
-        payload = {"form": form}
-        return render(request, "registry/address_list_csv_import_form.html", payload)
